Make the dev server use Bedrock core's template.
This commit changes where Bedrock development server is looking for
templates. Instead of using the local core/ directoy, it directs Bedrock
to its own core/ directory.

In addition, to allow styleguide templates to live in the local project,
this commit adds an underscore prefix to the _styleguide/ directory.

(The underscore is used to make sure those .pug files are not part of
the main content processed by the static build.)