Remove the development doc.
As the previous commit, this was too specific to how Mono worked with
Smart (e.g. intent to publish a package or where to fill an issue).

There was also some interesting consideration about BEM/ITCSS but this
can be found elsewhere anyway.

Should be a separate commit, but now links are blue #0000FF.